project_structure_new/requirements.txt
# Medical Report Simplifier - Root Dependencies
# This file includes all dependencies for the entire project

# ===== BACKEND DEPENDENCIES =====
# Flask Web Framework and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Security and Authentication
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.4

# Web Server
Werkzeug==2.3.7
gunicorn==21.2.0

# Template Engine
Jinja2==3.1.2
MarkupSafe==2.1.3

# ===== MACHINE LEARNING DEPENDENCIES =====
# Core ML and Data Science Libraries
torch>=1.9.0
transformers>=4.20.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Natural Language Processing
nltk>=3.7
spacy>=3.4.0
regex>=2022.7.0
scispacy>=0.5.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Streamlit for ML Interface
streamlit>=1.25.0
streamlit-option-menu>=0.3.2

# ===== SHARED DEPENDENCIES =====
# PDF and Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.7.0
python-docx>=0.8.11

# Image Processing
Pillow>=9.0.0
opencv-python>=4.6.0

# Utilities
click==8.1.7
itsdangerous==2.1.2
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
tqdm>=4.64.0

# Data Validation
marshmallow>=3.19.0

# Logging
structlog>=22.0.0

# Model Serialization
joblib>=1.2.0

# Statistical Analysis
statsmodels>=0.13.0

# ===== DEVELOPMENT DEPENDENCIES =====
# Testing
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Code Quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
